Add files via upload
Coffee Cup is into sale of beverages with focus on variants of coffee and tea. It has branches spread across all provinces and territories of Canada. With 102 stores, the profit or return on investment for shareholders is not encouraging. In order to develop growth strategy for profit maximization, sales data across all 102 stores were collected for 2012, 2018, and 2019. Coffee Cup’s goal is to be the neighborhood coffee house you know and love, and you think of when you desire a freshly brewed cup of hot coffee or desire a specialty tea selection. 
In this analysis, some of the performance metrics the management of Coffee Cup wants to look at include Cost of Goods, Store Performance, Customer Loyalty, Profits and Sales, Product Inventory, and any other metrics that the data analyst might want to consider. 
Coffee Cup’s corporate values include to earn the trust and loyalty of their customers. In achieving this, they want to: